requirements.txt
django>=3.2,<3.3
gunicorn
psycopg2-binary
django-dotenv
requests
django-storages
boto3
Pillow
BeautifulSoup4
djangorestframework
celery